pyproject.toml
[project]
name = "finance-agent"
version = "0.1.0"
description = "Personal finance AI agent — Gemini + MySQL + Streamlit"
requires-python = ">=3.11"

dependencies = [
    "google-genai>=0.8.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "pydantic-settings>=2.0.0",
    "streamlit>=1.32.0",
    "plotly>=5.20.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "pdf2image>=1.17.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agent", "db", "ui"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
